✍️ Update and improve project READMEs (#1223)
Our `README`s are pretty woeful and _very_ under maintained. This PR gives them some much needed care and attention and aims to make them actually useful for new starters to get going but also as a reference for the whole team when it comes to running and building the different parts of our platform.